Fix: do not send confirmations to message websocket
Problem: websocket users report messages appearing twice, as they
are received on both the P2P and IPFS pubsub topics.

Solution: ignore confirmations. The scaffolding was already in
place, we just need to return a `ProcessedMessage` instead of
a raw `MessageDb` from `MessageHandler.process()`.
