+++ tests/message_processing/test_process_aggregates.py
@@ -14,12 +14,12 @@ from aleph.db.accessors.aggregates import get_aggregate_by_key, get_aggregate_el
 from aleph.db.models import PendingMessageDb, MessageDb, AggregateElementDb, AggregateDb
 from aleph.handlers.content.aggregate import AggregateMessageHandler
 from aleph.handlers.message_handler import MessageHandler
-from aleph.jobs.job_utils import ProcessedMessage
 from aleph.jobs.process_pending_messages import PendingMessageProcessor
 from aleph.storage import StorageService
 from aleph.toolkit.timestamp import timestamp_to_datetime
 from aleph.types.channel import Channel
 from aleph.types.db_session import DbSessionFactory, DbSession
+from aleph.types.message_processing_result import ProcessedMessage
 from message_test_helpers import process_pending_messages
